Update state tool warning to say beta and only occur on activation
