+++ locale/en-us.yaml
@@ -813,7 +813,7 @@ err_os_not_a_directory:
 scripts_no_scripts:
   other: No scripts in your 'activestate.yaml' or no scripts to run in for your environment
 unstable_version_warning:
-  other: The State Tool is currently in a preview period, we are actively changing and adding features based on developer feedback.
+  other: The State Tool is currently in beta, we are actively changing and adding features based on developer feedback.
 export_cmd_description:
   other: Print information based on the provided subcommand
 export_recipe_cmd_description:
