+++ state/activate/testdata/activestate.yaml
@@ -31,7 +31,7 @@ languages:
         version: "2.*"
         build:
           override: --foo --bar --debug $variable.DEBUG --libDir $variable.PYTHONPATH --libc $platform.libc
-variables:
+constants:
   - name: DEBUG
     value: true
   - name: PYTHONPATH
@@ -40,7 +40,7 @@ variables:
         environment: dev,qa
   - name: PYTHONPATH
     value: '%projectDir%/src:%projectDir%/tests'
-hooks:
+events:
   - name: FIRST_INSTALL
     value: '%pythonExe% %projectDir%/setup.py prepare'
   - name: AFTER_UPDATE
